+++ osquery/core/watcher.h
@@ -210,11 +210,11 @@ class Watcher : private boost::noncopyable {
     worker_restarts_++;
   }
 
-  void workerStartTime(size_t start_time) {
+  void workerStartTime(uint64_t start_time) {
     worker_start_time_ = start_time;
   }
 
-  size_t workerStartTime() {
+  uint64_t workerStartTime() {
     return worker_start_time_;
   }
 
@@ -230,7 +230,7 @@ class Watcher : private boost::noncopyable {
   std::shared_ptr<PlatformProcess> worker_;
 
   /// Time the worker was started.
-  size_t worker_start_time_{0};
+  uint64_t worker_start_time_{0};
 
   /// Number of worker restarts NOT induced by a watchdog process.
   size_t worker_restarts_{0};
